[Bronze II] Title: 숫자, Time: 16 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -14,8 +14,8 @@ int main() {
         cout << 0;
         return 0;
     }
-    int s = min(a, b);
-    int e = max(a, b);
+    long long int s = min(a, b);
+    long long int e = max(a, b);
     cout << e - s - 1 << '\n';
     for (long long int i = s + 1; i < e; i++)
     {
